fix(dashboard): seperate router to ensure there is no circular dependency
